Fixes post labels being called for each request
When a unknown label is encountered, the label is created in the API. However, the call to loadLabels will never update the local cache, due to it checking the local cache before attempting to load the labels. This fix purges the 'labels' key from cache before calling loadLabels, preventing a new request for each load until the local cache is cleared.