Fix postLabels trying to create labels that exists
We previously tried to fix this issue because cache would not be
invalidated and therefore it would try to post the same label if it
occurred more than once per page.

This however was not sufficient because of another issue.
We checked if a value for the current language was present, not if the
key was present in and out of itself.

Unless a value was set for the language, the code would be null.
This fix fixes that